[IMP] fitness_app: improves code by adding a CheckAuthroization function that can be called in PrivateRouting and may be used for later components.
